MDL-57323 grades: Optionally hide grader name
Creates a new setting that enables users to hide the name of the grader. The site admin can control this setting's default value at gradeitemsettings. Course administrators can control this setting in the 'Grade item settings' section on /grade/edit/settings/index.php.

If enabled, the grader name will only be shown to users with the mod/assign:grade capability.